+++ lightwood/encoders/time_series/time_series.py
@@ -1,3 +1,4 @@
+from lightwood.encoders.time_series.helpers.common import *
 from lightwood.encoders.time_series.helpers.rnn_helpers import *
 from lightwood.encoders.time_series.helpers.transformer_helpers import *
 from lightwood.helpers.torch import LightwoodAutocast
