+++ tests/unit_tests/encoders/time_series/test_timeseries_rnn.py
@@ -5,34 +5,6 @@ from lightwood.encoders.time_series.helpers.rnn_helpers import *
 
 
 class TestRnnEncoder(unittest.TestCase):
-
-    def test_minmax_normalizer(self):
-        data = [[-100.0, -5.0, 0.0, 5.0, 100.0],
-                [-1000.0, -50.0, 0.0, 50.0, 1000.0],
-                [-500.0, -405.0, -400.0, -395.0, -300.0],
-                [300.0, 395.0, 400.0, 405.0, 500.0],
-                [0.0, 1e3, 1e6, 1e9, 1e12]]
-        normalizer = MinMaxNormalizer()
-        normalizer.prepare(data)
-        reconstructed = normalizer.decode(normalizer.encode(data))
-        self.assertTrue(np.allclose(data, reconstructed, atol=0.1))
-
-    def test_cat_normalizer(self):
-        data = [['a', 'b', 'c'],
-                ['c', 'b', 'b'],
-                ['a', 'a', None]]
-        encoded_target = [[[0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]],
-                          [[0, 0, 0, 1], [0, 0, 1, 0], [0, 0, 1, 0]],
-                          [[0, 1, 0, 0], [0, 1, 0, 0], [1, 0, 0, 0]]]
-        normalizer = CatNormalizer()
-        normalizer.prepare(data)
-        encoded = normalizer.encode(data)
-        self.assertTrue(encoded_target, encoded)
-        dec = normalizer.decode(encoded)
-        self.assertTrue(dec[-1][-1] == normalizer.unk)
-        dec[-1][-1] = None
-        self.assertTrue(data == dec)
-
     def test_overfit(self):
         series = [[1, 2, 3, 4, 5, 6],
                   [2, 3, 4, 5, 6, 7],
@@ -40,7 +12,7 @@ class TestRnnEncoder(unittest.TestCase):
                   [4, 5, 6, 7, 8, 9]]
 
         example = ([[1, 2, 3], [2, 3, 4], [3, 4, 5], [4, 5, 6]],  # query
-                   [4, 5, 6, 7])                                    # answer
+                   [4, 5, 6, 7])                                  # answer
 
         data = series * 100
         timesteps = 6
